Refactoring the parse structure
